refactor(componets): rearage componensts in folders
