+++ src/components/shared/movie-card/movie-card.jsx
