+++ src/components/views/login-view/login-view.jsx
@@ -9,8 +9,8 @@ import {
     Card,
 } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
-import logo from '../../assets/cinematix logo.svg';
-import { MessageModal } from '../message-modal/message-modal';
+import logo from '../../../assets/cinematix logo.svg';
+import { MessageModal } from '../../shared/message-modal/message-modal';
 
 export const LoginView = ({ onLoggedIn }) => {
     const [username, setUsername] = useState('');
